move code to decorator
